phtread: Change stacksize type from long to size_t
follow the standard prototype:
https://pubs.opengroup.org/onlinepubs/9699919799/functions/pthread_attr_getstack.html

+++ libs/libc/pthread/pthread_attr_getstack.c
@@ -49,7 +49,7 @@
  ****************************************************************************/
 
 int pthread_attr_getstack(FAR pthread_attr_t *attr,
-                          FAR void **stackaddr, FAR long *stacksize)
+                          FAR void **stackaddr, FAR size_t *stacksize)
 {
   int ret;
